lunrclient.__init__: empty
This moves all class and function definitions from lunrclient.__init__
to:

 - base
 - storage
 - lunr

All references to affected classes and functions are updated
accordingly.
